requirements.txt
requests
pygments
beautifulsoup4
tiktoken
tqdm
markdown
python-docx
reportlab
